@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { User } from '../../shared/interfaces';
 import { AuthService } from '../shared/services/auth.servise';
-import { Router } from '@angular/router';
+import {ActivatedRoute, Params, Router} from '@angular/router';
 
 @Component({
   selector: 'app-login-page',
@@ -13,10 +13,20 @@ export class LoginPageComponent implements OnInit {
 
   form: FormGroup;
   submitted = false;
+  message: string;
 
-  constructor(private auth: AuthService, private router: Router) { }
+  constructor(
+    public auth: AuthService,
+    private router: Router,
+    private route: ActivatedRoute
+  ) {}
 
   ngOnInit() {
+    this.route.queryParams.subscribe((params: Params) => {
+      if (params['loginAgain']) {
+        this.message = 'Будь ласка введіть дані!';
+      }
+    });
     this.form = new FormGroup({
       email: new FormControl(null, [
         Validators.required,
@@ -41,6 +51,8 @@ export class LoginPageComponent implements OnInit {
     this.auth.login(user).subscribe(() => {
       this.form.reset();
       this.router.navigate(['/admin', 'dashboard']);
+      this.submitted = false;
+    }, () => {
       this.submitted = false;
     });
   }
